+++ src/app/builder/[buildId]/page.tsx
@@ -20,13 +20,13 @@ import {
 } from '../actions'
 import { toast } from 'react-toastify'
 import { useRouter } from 'next/navigation'
-import { ExtendedBuild } from '../types'
 import {
   buildStateToCsvData,
   buildStateToMasonryItems,
   extendedBuildToBuildState,
-} from '../utils'
+} from '../../(lib)/build'
 import { cn } from '@/app/(lib)/utils'
+import { ExtendedBuild } from '@/app/(types)/build'
 
 export default function Page({
   params: { build },