+++ src/app/profile/favorited-builds/page.tsx
@@ -3,8 +3,8 @@ import { prisma } from '@/app/(lib)/db'
 import ViewBuildButton from '../(components)/ViewBuildButton'
 import CopyBuildUrlButton from '../(components)/CopyBuildUrlButton'
 import { StarIcon } from '@heroicons/react/24/solid'
-import { ExtendedBuild } from '@/app/builder/types'
-import { extendedBuildToBuildState } from '@/app/builder/utils'
+import { extendedBuildToBuildState } from '@/app/(lib)/build'
+import { ExtendedBuild } from '@/app/(types)/build'
 
 async function getBuilds() {
   const session = await getServerSession()